Move Rc cursor impls to cursor/mod.rs
